requirements.txt
beautifulsoup4==4.12.3
faiss-cpu==1.8.0.post1
huggingface-hub==0.25.1
langchain==0.3.2
langchain-community=0.3.1
langchain-core==0.3.8
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.1.1
streamlit==1.39.0
tokenizers==0.20.0
torch==2.2.2
transformers==4.45.1
uuid6==2024.7.10